refactor: split monolithic index.ts into modular file structure
Extract the single 846-line src/index.ts into focused modules:
- src/types.ts: 7 data interfaces
- src/moodle-client.ts: axios factory, getCourseId, validateArrayResponse
- src/tools/definitions.ts: 9 tool schemas
- src/tools/courses.ts: listCourses, getCourseContents
- src/tools/students.ts: getStudents
- src/tools/assignments.ts: getAssignments, getSubmissions, provideFeedback, getSubmissionContent
- src/tools/quizzes.ts: getQuizzes, getQuizGrade
- src/server.ts: MoodleMcpServer class with thin router
- src/index.ts: 5-line entry point

Zero behavioral changes. All 25 existing tests pass without modification.

Relates to #7
